Improve CVE Yaml creation
- Simplify, ask for subdirectory
- Ensure a patch note is present when intervention is required
- Rename rake tasks
