+++ lib/cve_list.rb
@@ -1,257 +1,9 @@
 # frozen_string_literal: true
 
 module CveList
-  # we don't support configurations for versions 2.3.18.1 .. 2.3.18.13
-  # 2.3.18.14 is the oldest version available via our gemserver
-  LTS_VERSIONS_2_3 = [
-    '2.3.18.14',
-    '2.3.18.15',
-    '2.3.18.16',
-    '2.3.18.17',
-    '2.3.18.18',
-    '2.3.18.19',
-    '2.3.18.20',
-    '2.3.18.21',
-    '2.3.18.22',
-    '2.3.18.23',
-    '2.3.18.24',
-    '2.3.18.25',
-    '2.3.18.26',
-    '2.3.18.27',
-    '2.3.18.28',
-    '2.3.18.29',
-    '2.3.18.30',
-    '2.3.18.31',
-    '2.3.18.32',
-    '2.3.18.33',
-    '2.3.18.34',
-    '2.3.18.35',
-    '2.3.18.36',
-    '2.3.18.37',
-    '2.3.18.38',
-    '2.3.18.39',
-    '2.3.18.40',
-    '2.3.18.41',
-    '2.3.18.42',
-    '2.3.18.43',
-    '2.3.18.44',
-    '2.3.18.45',
-    '2.3.18.46',
-    '2.3.18.47',
-    '2.3.18.48',
-    '2.3.18.49',
-    '2.3.18.50',
-    '2.3.18.51',
-    '2.3.18.52',
-    '2.3.18.53',
-    '2.3.18.54',
-    '2.3.18.55',
-    '2.3.18.56',
-    '2.3.18.57',
-    '2.3.18.58',
-    '2.3.18.59',
-  ].freeze
-
-  LTS_VERSIONS_3_2 = [
-    '3.2.21.1',
-    '3.2.22.1',
-    '3.2.22.2',
-    '3.2.22.3',
-    '3.2.22.4',
-    '3.2.22.5',
-    '3.2.22.6',
-    '3.2.22.7',
-    '3.2.22.8',
-    '3.2.22.9',
-    '3.2.22.10',
-    '3.2.22.11',
-    '3.2.22.12',
-    '3.2.22.13',
-    '3.2.22.14',
-    '3.2.22.15',
-    '3.2.22.16',
-    '3.2.22.17',
-    '3.2.22.18',
-    '3.2.22.19',
-    '3.2.22.20',
-    '3.2.22.21',
-    '3.2.22.22',
-    '3.2.22.23',
-    '3.2.22.24',
-    '3.2.22.25',
-    '3.2.22.26',
-    '3.2.22.27',
-    '3.2.22.28',
-    '3.2.22.29',
-    '3.2.22.30',
-    '3.2.22.31',
-    '3.2.22.32',
-    '3.2.22.33',
-    '3.2.22.34',
-    '3.2.22.35',
-    '3.2.22.36',
-    '3.2.22.37',
-    '3.2.22.38',
-    '3.2.22.39',
-    '3.2.22.40',
-    '3.2.22.41',
-    '3.2.22.42',
-    '3.2.22.43',
-    '3.2.22.44',
-    '3.2.22.45',
-    '3.2.22.46',
-    '3.2.22.47',
-    '3.2.22.48',
-    '3.2.22.49',
-  ].freeze
-
-  LTS_VERSIONS_4_2 = [
-    '4.2.10.0',
-    '4.2.10.1',
-    '4.2.11.10',
-    '4.2.11.11',
-    '4.2.11.12',
-    '4.2.11.13',
-    '4.2.11.14',
-    '4.2.11.15',
-    '4.2.11.16',
-    '4.2.11.17',
-    '4.2.11.18',
-    '4.2.11.19',
-    '4.2.11.20',
-    '4.2.11.21',
-    '4.2.11.22',
-    '4.2.11.23',
-    '4.2.11.24',
-    '4.2.11.25',
-    '4.2.11.26',
-    '4.2.11.27',
-    '4.2.11.28',
-    '4.2.11.29',
-    '4.2.11.30',
-    '4.2.11.31',
-    '4.2.11.32',
-    '4.2.11.33',
-    '4.2.11.34',
-    '4.2.11.35',
-    '4.2.11.36',
-    '4.2.11.37',
-    '4.2.11.38',
-    '4.2.11.39',
-  ].freeze
-
-  LTS_VERSIONS_5_2 = [
-    '5.2.6.10',
-    '5.2.6.11',
-    '5.2.6.12',
-    '5.2.7.10',
-    '5.2.7.11',
-    '5.2.8.10',
-    '5.2.8.11',
-    '5.2.8.12',
-    '5.2.8.13',
-    '5.2.8.14',
-    '5.2.8.15',
-    '5.2.8.16',
-    '5.2.8.17',
-    '5.2.8.18',
-    '5.2.8.19',
-    '5.2.8.20',
-    '5.2.8.21',
-    '5.2.8.22',
-    '5.2.8.23',
-    '5.2.8.24',
-    '5.2.8.25',
-    '5.2.8.26',
-    '5.2.8.27',
-    '5.2.8.28',
-    '5.2.8.29',
-    '5.2.8.30',
-    '5.2.8.31',
-    '5.2.8.32',
-    '5.2.8.34',
-  ].freeze
-
-  LTS_VERSIONS_6_1 = [
-    '6.1.7.18',
-    '6.1.7.19',
-    '6.1.7.20',
-    '6.1.7.21',
-    '6.1.7.22',
-    '6.1.7.23',
-    '6.1.7.24',
-    '6.1.7.25',
-    '6.1.7.26',
-    '6.1.7.27',
-    '6.1.7.28',
-    '6.1.7.30',
-  ].freeze
-
-  LTS_VERSIONS = [
-    *LTS_VERSIONS_2_3,
-    *LTS_VERSIONS_3_2,
-    *LTS_VERSIONS_4_2,
-    *LTS_VERSIONS_5_2,
-    *LTS_VERSIONS_6_1,
-  ].freeze
-
-  RACK_VERSIONS_1_4 = [
-    '1.4.7.10',
-    '1.4.7.11',
-    '1.4.7.12',
-    '1.4.7.13',
-    '1.4.7.14',
-    '1.4.7.15',
-    '1.4.7.16',
-    '1.4.7.17',
-    '1.4.7.18',
-    '1.4.7.19',
-    '1.4.7.20',
-    '1.4.7.21',
-    '1.4.7.22',
-    '1.4.7.23',
-    '1.4.7.24',
-  ].freeze
-
-  RACK_VERSIONS_1_6 = [
-    '1.6.13.10',
-    '1.6.13.11',
-    '1.6.13.12',
-    '1.6.13.13',
-    '1.6.13.14',
-    '1.6.13.15',
-    '1.6.13.16',
-    '1.6.13.17',
-    '1.6.13.18',
-    '1.6.13.19',
-    '1.6.13.20',
-    '1.6.13.21',
-    '1.6.13.22',
-  ].freeze
-
-  RACK_VERSIONS_2_2 = [
-    '2.2.4.10',
-    '2.2.4.11',
-    '2.2.4.12',
-    '2.2.4.13',
-    '2.2.6.14',
-    '2.2.8.10',
-    '2.2.9.10',
-    '2.2.11.10',
-    '2.2.12.10',
-    '2.2.13.10',
-    '2.2.14.10',
-    '2.2.17.10',
-    '2.2.18.10',
-  ].freeze
-
-  RACK_VERSIONS = [
-    *RACK_VERSIONS_1_4,
-    *RACK_VERSIONS_1_6,
-    *RACK_VERSIONS_2_2,
-  ].freeze
 end
 
+require_relative 'cve_list/versions'
 require_relative 'cve_list/input/parser'
 require_relative 'cve_list/output_generator/snyk_config'
 require_relative 'cve_list/output_generator/bundler_audit_config'
